# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Business": {
        "properties": {
          "enterprise_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enterprise Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "receiving_processes": {
            "items": {
              "type": "string"
            },
            "title": "Receiving Processes",
            "type": "array"
          },
          "roles": {
            "items": {
              "enum": [
                "buyer",
                "seller"
              ],
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "roles",
          "receiving_processes",
          "location"
        ],
        "title": "Business",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Listing": {
        "description": "Public listing. Deliberately has no seller floor field.",
        "properties": {
          "asking_price_paise_per_tonne": {
            "description": "INR paise per tonne",
            "minimum": 0.0,
            "title": "Asking Price Paise Per Tonne",
            "type": "integer"
          },
          "available_quantity_kg": {
            "description": "kilograms",
            "minimum": 0.0,
            "title": "Available Quantity Kg",
            "type": "integer"
          },
          "contamination_notes": {
            "title": "Contamination Notes",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "moisture_pct": {
            "title": "Moisture Pct",
            "type": "number"
          },
          "pickup_window": {
            "$ref": "#/components/schemas/Window"
          },
          "seller": {
            "$ref": "#/components/schemas/Business"
          },
          "status": {
            "enum": [
              "open",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "seller",
          "material_id",
          "available_quantity_kg",
          "asking_price_paise_per_tonne",
          "moisture_pct",
          "contamination_notes",
          "pickup_window",
          "location",
          "status"
        ],
        "title": "Listing",
        "type": "object"
      },
      "ListingCreate": {
        "properties": {
          "asking_price_paise_per_tonne": {
            "description": "INR paise per tonne",
            "minimum": 0.0,
            "title": "Asking Price Paise Per Tonne",
            "type": "integer"
          },
          "available_quantity_kg": {
            "description": "kilograms",
            "exclusiveMinimum": 0.0,
            "title": "Available Quantity Kg",
            "type": "integer"
          },
          "contamination_notes": {
            "default": "",
            "title": "Contamination Notes",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "moisture_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Moisture Pct",
            "type": "number"
          },
          "pickup_window": {
            "$ref": "#/components/schemas/Window"
          },
          "seller_floor_paise_per_tonne": {
            "description": "INR paise per tonne",
            "minimum": 0.0,
            "title": "Seller Floor Paise Per Tonne",
            "type": "integer"
          }
        },
        "required": [
          "material_id",
          "available_quantity_kg",
          "asking_price_paise_per_tonne",
          "seller_floor_paise_per_tonne",
          "moisture_pct",
          "pickup_window"
        ],
        "title": "ListingCreate",
        "type": "object"
      },
      "ListingUpdate": {
        "properties": {
          "asking_price_paise_per_tonne": {
            "anyOf": [
              {
                "description": "INR paise per tonne",
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asking Price Paise Per Tonne"
          },
          "pickup_window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Window"
              },
              {
                "type": "null"
              }
            ]
          },
          "seller_floor_paise_per_tonne": {
            "anyOf": [
              {
                "description": "INR paise per tonne",
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Floor Paise Per Tonne"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "open",
                  "closed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "ListingUpdate",
        "type": "object"
      },
      "Location": {
        "properties": {
          "district": {
            "title": "District",
            "type": "string"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "precision": {
            "default": "district",
            "enum": [
              "district",
              "confirmed_site"
            ],
            "title": "Precision",
            "type": "string"
          }
        },
        "required": [
          "district"
        ],
        "title": "Location",
        "type": "object"
      },
...
      "OwnerListing": {
        "properties": {
          "asking_price_paise_per_tonne": {
            "description": "INR paise per tonne",
            "minimum": 0.0,
            "title": "Asking Price Paise Per Tonne",
            "type": "integer"
          },
          "available_quantity_kg": {
            "description": "kilograms",
            "minimum": 0.0,
            "title": "Available Quantity Kg",
            "type": "integer"
          },
          "contamination_notes": {
            "title": "Contamination Notes",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "moisture_pct": {
            "title": "Moisture Pct",
            "type": "number"
          },
          "pickup_window": {
            "$ref": "#/components/schemas/Window"
          },
          "seller": {
            "$ref": "#/components/schemas/Business"
          },
          "seller_floor_paise_per_tonne": {
            "description": "INR paise per tonne",
            "minimum": 0.0,
            "title": "Seller Floor Paise Per Tonne",
            "type": "integer"
          },
          "status": {
            "enum": [
              "open",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "seller",
          "material_id",
          "available_quantity_kg",
          "asking_price_paise_per_tonne",
          "moisture_pct",
          "contamination_notes",
          "pickup_window",
          "location",
          "status",
          "seller_floor_paise_per_tonne"
        ],
        "title": "OwnerListing",
        "type": "object"
      },
...
      "Page_Listing_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Listing"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[Listing]",
        "type": "object"
      },
...
      "Page_OwnerListing_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OwnerListing"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[OwnerListing]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Window": {
        "properties": {
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "Window",
        "type": "object"
      }
    },
    "securitySchemes": {
      "AccessToken": {
        "description": "AUTH_MODE=dev: paste `dev:<email>` (for example `dev:buyer1@demo.bitnbuild.local`). AUTH_MODE=supabase: paste a Supabase access token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "B2B exchange for secondary industrial materials. Prices are entered by participants; freight charges are entered or configured estimates. No market price feed, benchmark, or carrier booking is involved.",
    "title": "Karnataka Industrial Byproduct Exchange",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/listings": {
      "get": {
        "operationId": "list_listings_api_v1_listings_get",
        "parameters": [
          {
            "in": "query",
            "name": "material_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Material Id"
            }
          },
          {
            "in": "query",
            "name": "district",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Listing_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "List Listings",
        "tags": [
          "listings"
        ]
      },
      "post": {
        "operationId": "create_listing_api_v1_listings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerListing"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Create Listing",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/v1/listings/{listing_id}": {
      "get": {
        "operationId": "get_listing_api_v1_listings__listing_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Listing"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Get Listing",
        "tags": [
          "listings"
        ]
      },
      "patch": {
        "operationId": "update_listing_api_v1_listings__listing_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerListing"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Update Listing",
        "tags": [
          "listings"
        ]
      }
    },
...
    "/api/v1/me/listings": {
      "get": {
        "operationId": "my_listings_api_v1_me_listings_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_OwnerListing_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "My Listings",
        "tags": [
          "listings"
        ]
      }
    },
...
  }
}